requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2.12
django-cpf==0.1.0
django-phonenumber-field==6.1.0
phonenumbers==8.12.45
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
